Move CompressionOptions to core, remove server dep from netty-core
